tests/wpt/web-platform-tests/html/semantics/document-metadata/the-base-element/base_href_specified.sub.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: base_href_specified</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#the-base-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<base id="base" href="http://{{domains[www]}}:{{ports[http][0]}}">
<div id="log"></div>
<img id="test" src="test.ico" style="display:none">

<script>
  var testElement;
  var baseElement;

  setup(function() {
    testElement = document.getElementById("test");
    baseElement = document.getElementById("base");
  });

  test(function() {
    assert_equals(baseElement.href, "http://{{domains[www]}}:{{ports[http][0]}}/", "The href attribute of the base element is incorrect.");
  }, "The href attribute of the base element is specified");

  test(function() {
    assert_equals(testElement.src, "http://{{domains[www]}}:{{ports[http][0]}}/test.ico", "The src attribute of the img element is incorrect.");
  }, "The src attribute of the img element must relative to the href attribute of the base element");
</script>
